Disk-based construction improvements (#125)
* First

* Additional tests for SortedSetDisk-based graph construction

* Increased buffer using a dynmically allocated array

* Add test for correctly decrementing from the end

* First benchmark for file merging

* Use a vector based heap, as it's faster

* Rename the benchmark file

* Benchmark for file pairs

* Switch back to profiling file

* Allow only distinct elements in heap

* Allow back duplicate  elements in heap

* Updated comment

* minor

* Merge duplicates when popping from heap

* Allow setting temp directory. Randomize temp directory to avoid conflicts. ChunkedWaitQueue benchmark

* Addressed review comments

* Use mkdtemp instead of hand-crafted temp dir

* Add back missing includes

* Use an optional instead of bool+value to avoid spurious warnings in gcc

* Add missing include

* Invert filter suffix condition

* Minor comments change

* Fix log message

* Smarter benchmark, Fixed tmp_dir

* typo

* Addressed comments

* Use tmpfile in both benchmarks

* Slightly more efficient benchmark

* Fix message

* Unify the 2 heaps

* Make parameters identical to previous runs

* Revert "Make parameters identical to previous runs"

This reverts commit 3de32c87b70f44a469d8644c8df6efffd1edb80e.

* Revert "Unify the 2 heaps"

This reverts commit 87455e8982c2b5cb5036a6975c41b95cefd4bf8e.

* Latest round of review comments

* Updated comment

* Changed params

* Fixed typo in test_boss_construct

* Make cleanup optional.

* Moved new benchmarks to benchmkars/common

* Fix comment

* Fix init in kmer collector

* minor fixes and cleanups

* generalized merge heap, fixed bug with count overflow

* cleanup

* slightly optimized the performance: don't use std::optional

* comments fixed

* Test that overflows are handled correctly

* Test for in-memory de-dupe overflows

* Clarifying comment in test

* Update tclap

+++ metagraph/src/common/file_merger.hpp
@@ -12,107 +12,35 @@
 
 namespace mg {
 namespace common {
-/**
- * Given a list of n source files, containing ordered elements of type T, merge the n
- * sources into a single (ordered) list and delete the original files.
- * @param sources the files containing sorted lists of type T
- * @param on_new_item callback to invoke when a new element was merged
- *
- * @return the total number of elements read from all files
- *
- * Note: this method blocks until all the data was successfully merged.
- */
-template <typename T>
-uint64_t merge_files(const std::vector<std::string> sources,
-                     std::function<void(const T &)> on_new_item) {
-    // start merging disk chunks by using a heap to store the current element
-    // from each chunk
-    std::vector<std::ifstream> chunk_files(sources.size());
-    uint64_t num_elements = 0;
-
-    std::priority_queue<std::pair<T, uint32_t>,
-                        std::vector<std::pair<T, uint32_t>>,
-                        utils::GreaterFirst>
-                    merge_heap;
-    T data_item;
-    for (uint32_t i = 0; i < sources.size(); ++i) {
-        // TODO: use a bigger buffer in stream to reduce the number
-        // of random reads or insert into the heap a batch of elements at a time
-        chunk_files[i].open(sources[i], std::ios::in | std::ios::binary);
-        if (!chunk_files[i].good()) {
-            logger->error("Error: Unable to open chunk file '{}'", sources[i]);
-            std::exit(EXIT_FAILURE);
-        }
-        if (chunk_files[i].read(reinterpret_cast<char *>(&data_item), sizeof(T))) {
-            merge_heap.emplace(data_item, i);
-            num_elements++;
-        }
-    }
-
-    // initialized to suppress maybe-uninitialized warnings in GCC
-    T last_written = {};
-
-    bool has_written = false;
-
-    while (!merge_heap.empty()) {
-        std::pair<T, uint32_t> smallest = merge_heap.top();
-        merge_heap.pop();
-
-        if (!has_written || smallest.first != last_written) {
-            has_written = true;
-            on_new_item(smallest.first);
-            last_written = smallest.first;
-        }
-
-        if (chunk_files[smallest.second]
-             && chunk_files[smallest.second].read(reinterpret_cast<char *>(&data_item),
-                                                  sizeof(T))) {
-            merge_heap.emplace(data_item, smallest.second);
-            num_elements++;
-        }
-    }
-
-    std::for_each(sources.begin(), sources.end(),
-                  [](const std::string &s) { std::filesystem::remove(s); });
-
-    return num_elements;
-}
 
 /**
- * A heap that merges elements that are equal by adding their counts.
+ * Heap implemented as a sorted vector.
+ *
  * The heap uses a vector as the underlying data structure, thus making it efficient
- * only for a small (<100) number of elements.
- * //TODO(ddanciu): profile is using an actual binary heap is faster in practice
+ * only for a small (<1000) number of elements.
  * @tparam T the actual heap element type; this is the type tested for equality
- * @tparam C the count type for elements in the heap (typically some unsigned integer)
  */
- //TODO(ddanciu) - instead of using MergingHeap, we can use a simple
- // std::priority_queue which stores duplicate elements and merge elements when popping
- // Profile which one is faster. Also profile if using an  std::vector as in MergeQueue
- // is faster than using an std::queue.
-template <typename T, typename C>
-class MergingHeap {
-    /** The heap stores triplets of the form <Element, Count, SourceIndex> */
-    using value_type = std::tuple<T, C, uint32_t> ;
+// Note: profiling shows that using a sorted vector instead of a std::priority queue is
+// faster up to ~1000 elements.  Using an unsorted vector (faster insert,
+// slower pop()) is ~40% slower. Preventing duplicates in the heap so that we don't
+// need to test for dupes at pop time is ~60% slower.
+// Note: Profiling indicates that merging elements within the heap is ~30% slower than
+// inserting duplicates and merging them when popping from the heap, as we do now
+template <typename T, class Compare = std::greater<T>>
+class MergeHeap {
+    /** The heap stores pairs <Element, SourceIndex> */
+    using value_type = std::pair<T, uint32_t>;
 
   public:
-    bool emplace(T el, C count, uint32_t idx) {
-        auto it = std::lower_bound(els.begin(), els.end(), value_type(el, count, idx),
-                                   [](const value_type &a, const value_type &b) {
-                                       return std::get<0>(a) > std::get<0>(b);
+    void emplace(T el, uint32_t idx) {
+        auto it = std::lower_bound(els.begin(), els.end(), el,
+                                   [this](const value_type &p, const T &v) {
+                                       return compare_(p.first, v);
                                    });
-        if (it != els.end() && el == std::get<0>(*it)) {
-            C &value = std::get<1>(*it);
-            if (value < std::numeric_limits<C>::max() - count) {
-                value += count;
-            } else {
-                value = std::numeric_limits<C>::max();
-            }
-            return true;
-        }
-        els.emplace(it, el, count, idx);
-        return false;
+        els.emplace(it, el, idx);
     }
+
+    const value_type& top() const { return els.back(); }
 
     value_type pop() {
         value_type result = els.back();
@@ -125,7 +53,81 @@ class MergingHeap {
   private:
     // elements stored in decreasing order of the first tuple member
     std::vector<value_type> els;
+    Compare compare_ = Compare();
 };
+
+
+/**
+ * Given a list of n source files, containing ordered elements of type T, merge the n
+ * sources into a single (ordered) list and delete the original files.
+ * @param sources the files containing sorted lists of type T
+ * @param on_new_item callback to invoke when a new element was merged
+ *
+ * @return the total number of elements read from all files
+ *
+ * Note: this method blocks until all the data was successfully merged.
+ */
+template <typename T>
+uint64_t merge_files(const std::vector<std::string> sources,
+                     std::function<void(const T &)> on_new_item,
+                     bool cleanup = true) {
+    // start merging disk chunks by using a heap to store the current element
+    // from each chunk
+    std::vector<std::ifstream> chunk_files(sources.size());
+    uint64_t num_elements_read = 0;
+
+    MergeHeap<T> merge_heap;
+    T data_item;
+    // profiling indicates that setting a larger buffer slightly increases performance
+    std::unique_ptr<char[]> buffer(new char[sources.size() * 1024 * 1024]);
+    for (uint32_t i = 0; i < sources.size(); ++i) {
+        chunk_files[i].rdbuf()->pubsetbuf((buffer.get() + i * 1024 * 1024), 1024 * 1024);
+        chunk_files[i].open(sources[i], std::ios::in | std::ios::binary);
+        if (!chunk_files[i].good()) {
+            logger->error("Unable to open chunk file '{}'", sources[i]);
+            std::exit(EXIT_FAILURE);
+        }
+        if (chunk_files[i].read(reinterpret_cast<char *>(&data_item), sizeof(T))) {
+            merge_heap.emplace(data_item, i);
+            num_elements_read++;
+        }
+    }
+
+    if (merge_heap.empty()) {
+        if (cleanup) {
+            std::for_each(sources.begin(), sources.end(),
+                          [](const std::string &s) { std::filesystem::remove(s); });
+        }
+        return num_elements_read;
+    }
+
+    T last_written = merge_heap.top().first;
+    on_new_item(last_written);
+
+    while (!merge_heap.empty()) {
+        auto [smallest, chunk_index] = merge_heap.pop();
+
+        if (smallest != last_written) {
+            on_new_item(smallest);
+            last_written = smallest;
+        }
+
+        if (chunk_files[chunk_index]
+            && chunk_files[chunk_index].read(reinterpret_cast<char *>(&data_item),
+                                             sizeof(T))) {
+            merge_heap.emplace(data_item, chunk_index);
+            num_elements_read++;
+        }
+    }
+
+    if (cleanup) {
+        std::for_each(sources.begin(), sources.end(),
+                      [](const std::string &s) { std::filesystem::remove(s); });
+    }
+
+    return num_elements_read;
+}
+
 
 /**
  * Given a list of n source files, containing ordered pairs of  <element, count>,
@@ -141,51 +143,68 @@ class MergingHeap {
  */
 template <typename T, typename C>
 uint64_t merge_files(const std::vector<std::string> sources,
-                     std::function<void(const std::pair<T, C> &)> on_new_item) {
-    // Convenience type for storing a pair and it's source file index
-    using CountedEl = std::tuple<T, C, uint32_t>;
+                     std::function<void(const std::pair<T, C> &)> on_new_item,
+                     bool cleanup = true) {
     // start merging disk chunks by using a heap to store the current element
     // from each chunk
     std::vector<std::ifstream> chunk_files(sources.size());
     uint64_t num_elements = 0;
 
-    MergingHeap<T, C> merge_heap;
+    MergeHeap<std::pair<T, C>, utils::GreaterFirst> merge_heap;
     std::pair<T, C> data_item;
+    std::unique_ptr<char[]> buffer(new char[sources.size() * 1024 * 1024]);
     for (uint32_t i = 0; i < sources.size(); ++i) {
-        // TODO: try setting a larger buffer size in ifstream to reduce disk I/O
+        chunk_files[i].rdbuf()->pubsetbuf((buffer.get() + i * 1024 * 1024), 1024 * 1024);
         chunk_files[i].open(sources[i], std::ios::in | std::ios::binary);
         if (!chunk_files[i].good()) {
             logger->error("Unable to open chunk file '{}'", sources[i]);
             std::exit(EXIT_FAILURE);
         }
-        bool found = true;
-        while (found) {
-            if (chunk_files[i].read(reinterpret_cast<char *>(&data_item),
-                                    sizeof(std::pair<T, C>))) {
-                found = merge_heap.emplace(data_item.first, data_item.second, i);
-                num_elements++;
-            } else {
-                found = false;
-            }
-        }
-    }
-
-    while (!merge_heap.empty()) {
-        CountedEl smallest = merge_heap.pop();
-        on_new_item({ std::get<0>(smallest), std::get<1>(smallest) });
-
-        bool found = true;
-        uint32_t chunk_index = std::get<2>(smallest);
-        while (found && chunk_files[chunk_index]
-               && chunk_files[chunk_index].read(reinterpret_cast<char *>(&data_item),
-                                                sizeof(std::pair<T, C>))) {
-            found = merge_heap.emplace(data_item.first, data_item.second, chunk_index);
+        if (chunk_files[i].read(reinterpret_cast<char *>(&data_item),
+                                sizeof(std::pair<T, C>))) {
+            merge_heap.emplace(data_item, i);
             num_elements++;
         }
     }
 
-    std::for_each(sources.begin(), sources.end(),
-                  [](const std::string &s) { std::filesystem::remove(s); });
+    if (merge_heap.empty()) {
+        if (cleanup) {
+            std::for_each(sources.begin(), sources.end(),
+                          [](const std::string &s) { std::filesystem::remove(s); });
+        }
+        return num_elements;
+    }
+
+    // initialize the smallest element
+    std::pair<T, C> current = { merge_heap.top().first.first, 0 };
+
+    while (!merge_heap.empty()) {
+        auto [smallest, chunk_index] = merge_heap.pop();
+
+        if (smallest.first != current.first) {
+            on_new_item(current);
+            current = smallest;
+        } else {
+            if (current.second < std::numeric_limits<C>::max() - smallest.second) {
+                current.second += smallest.second;
+            } else {
+                current.second = std::numeric_limits<C>::max();
+            }
+        }
+
+        if (chunk_files[chunk_index]
+            && chunk_files[chunk_index].read(reinterpret_cast<char *>(&data_item),
+                                             sizeof(std::pair<T, C>))) {
+            merge_heap.emplace(data_item, chunk_index);
+            num_elements++;
+        }
+    }
+    on_new_item(current);
+
+    if (cleanup) {
+        std::for_each(sources.begin(), sources.end(),
+                      [](const std::string &s) { std::filesystem::remove(s); });
+    }
 
     return num_elements;
 }
